% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backsolve.R
\name{backsolveBatch}
\alias{backsolveBatch}
\title{Backsolve function on GPU}
\usage{
backsolveBatch(
  C,
  A,
  B,
  numbatchB,
  diagIsOne,
  Nglobal,
  Nlocal,
  NlocalCache,
  Cstartend,
  Astartend,
  Bstartend,
  verbose = FALSE
)
}
\arguments{
\item{C}{a vclMatrix}

\item{A}{a vclMatrix, unit lower triangular}

\item{B}{a vclMatrix}

\item{numbatchB}{number of batches in B}

\item{diagIsOne}{logical, indicates if all the diagonal entries in A are one}
}
\value{
a vclMatrix C that satisfies A*C=B
}
\description{
Backsolve function on GPU
}
