% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cholBatch.R
\name{cholBatch}
\alias{cholBatch}
\title{Backsolve function on GPU}
\usage{
cholBatch(
  A,
  D,
  numbatchD,
  Nglobal,
  Nlocal,
  NlocalCache,
  Astartend,
  Dstartend,
  verbose = FALSE
)
}
\arguments{
\item{A}{a vclMatrix, positive definite}

\item{D}{a vclMatrix, diagonal elements of LDL^t}

\item{numbatchD}{number of batches in D}
}
\value{
vclMatrix L and D that satisfies A=LDL^T
}
\description{
Backsolve function on GPU
}
