% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossprodBatch.R
\name{crossprodBatch}
\alias{crossprodBatch}
\title{crossprodBatch function on GPU}
\usage{
crossprodBatch(
  C,
  A,
  D,
  invertD,
  Nglobal,
  Nlocal,
  NlocalCache,
  Cstartend,
  Astartend,
  Dstartend
)
}
\arguments{
\item{A}{rectangular matrices}

\item{D}{rectangular matrix, columns are diagonals}

\item{invertD}{set to 1 for C = t(A) D^(-1) A}

\item{Nglobal}{vector of number of global work items}

\item{Nlocal}{vector of number of local work items}

\item{NlocalCache}{elements in local cache}
}
\value{
C = A^T A or A^T D A or A^T D^(-1) A, output matrices, stacked row-wise
}
\description{
crossprodBatch function on GPU
}
