% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemmbatch.R
\name{gemmBatch}
\alias{gemmBatch}
\title{Gemmbatch on GPU}
\usage{
gemmBatch(
  A,
  B,
  Arowbatch,
  Browbatch,
  Acolbatch,
  Bcolbatch,
  need_transpose,
  workgroupSize,
  verbose = FALSE
)
}
\arguments{
\item{A}{a vclMatrix consists of batches of submatrices}

\item{B}{a vclMatrix consists of batches of submatrices}

\item{Acolbatch}{number of batches in the column of A}

\item{Bcolbatch}{number of batches in the column of B}

\item{rowbatch}{number of batches in row}
}
\value{
a vclMatrix of submatrices multiplication results
}
\description{
Gemmbatch on GPU
}
