% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gemmBatch2backend}
\alias{gemmBatch2backend}
\title{Multiplies a rectangular matrix by a rectangular matrix in batches}
\usage{
gemmBatch2backend(
  A,
  B,
  C,
  transposeABC,
  submatrixA,
  submatrixB,
  submatrixC,
  batches,
  workgroupSize,
  NlocalCache,
  verbose
)
}
\arguments{
\item{A}{rectangular matrices}

\item{B}{rectangular matrices}

\item{C}{output matrices, stacked row-wise}

\item{Nglobal}{vector of number of global work items//' @param Nlocal vector of number of local work items//' @param NlocalCache elements in local cache}
}
\description{
Multiplies a rectangular matrix by a rectangular matrix in batches
}
