% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likfit_givenBetas.R
\name{likfit_givenBeta}
\alias{likfit_givenBeta}
\title{Estimate Log-likelihood for Gaussian random fields when Betas are given}
\usage{
likfit_givenBeta(
  Betas,
  variances,
  Nobs,
  detVar,
  ssqY,
  XVYXVX,
  jacobian,
  form = c("loglik", "mlFixSigma"),
  minustwotimes = TRUE
)
}
\description{
Estimate Log-likelihood for Gaussian random fields when Betas are given
}
