% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likfitGpu.R
\name{likfitGpu}
\alias{likfitGpu}
\title{Estimate Log-likelihood for Gaussian random fields}
\usage{
likfitGpu(
  spatialmodel,
  type = c("double", "float"),
  paramsGpu,
  BoxCox,
  form = c("loglik", "ml", "mlFixSigma", "mlFixBeta", "reml", "remlPro"),
  NparamPerIter,
  nBetahats = 1,
  minustwotimes = TRUE,
  Nglobal,
  Nlocal,
  NlocalCache,
  betas = NULL,
  verbose = 1
)
}
\description{
Estimate Log-likelihood for Gaussian random fields
}
