% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maternParamsGpu.R
\name{maternGpuParam}
\alias{maternGpuParam}
\title{maternGpuParam}
\usage{
maternGpuParam(x, type = "double")
}
\arguments{
\item{x}{an R matrix of covariance parameters.}

\item{type}{precision type of parameters, either 'double' or 'float'.}
}
\value{
a `vclMatrix' object containing the GPU parameters.
}
\description{
Create parameters to be used by the `maternBatch' function on GPU.
}
