% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplyLowerDiagonalBatch.R
\name{multiplyLowerDiagonalBatch}
\alias{multiplyLowerDiagonalBatch}
\title{multiplyLowerDiagonalBatch function on GPU}
\usage{
multiplyLowerDiagonalBatch(
  output,
  L,
  D,
  B,
  diagIsOne,
  transformD,
  Nglobal,
  Nlocal,
  NlocalCache
)
}
\arguments{
\item{output}{the output = LDB}

\item{L}{lower triangular matrices in batce}

\item{D}{diagonal matrices in batch}

\item{B}{matrices in batch}

\item{diagIsOne}{logical, whether the diagonal of L is one}

\item{transformD}{how to transform D}

\item{Nglobal}{the size of the index space for use}

\item{Nlocal}{the work group size of the index space}

\item{NlocalCache}{a number}
}
\value{
the result matrices of LDB in batch
}
\description{
multiplyLowerDiagonalBatch function on GPU
}
