% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplyLowerDiagonalBatch.R
\name{multiplyLowerDiagonalBatch}
\alias{multiplyLowerDiagonalBatch}
\title{multiplyLowerDiagonalBatch}
\usage{
multiplyLowerDiagonalBatch(
  L,
  D,
  B,
  output,
  diagIsOne,
  transformD,
  Nglobal,
  Nlocal,
  NlocalCache
)
}
\arguments{
\item{L}{a `vclMatrix' consists of batches of lower triangular matrices}

\item{D}{a `vclMatrix' consists of diagonal matrices in batches, each row is one batch contains diagonal elements of D^T}

\item{B}{a `vclMatrix' consists of matrices in batches}

\item{output}{a `vclMatrix' which is the result of LD'B}

\item{diagIsOne}{a logical value indicating whether the diagonal of L is one}

\item{transformD}{a character string specifying how to transform D, can be any OpenCL C built-in math function}

\item{Nglobal}{the size of the index space for use}

\item{Nlocal}{the work group size of the index space}

\item{NlocalCache}{a integer number}
}
\value{
this function returns nothing, it modifies the input `vclMatrix' output in place.
}
\description{
Computes output = LD^T B in parallel batches on a GPU
}
\note{
the computed results are stored in `output'.
}
